+++ front/config.php
@@ -2,7 +2,10 @@
 
 include('../../../inc/includes.php');
 
-Session::checkRight(\GlpiPlugin\Ticketmigration\ProfileRight::RIGHT_CONFIG, READ);
+if (!Session::haveRight(\GlpiPlugin\Ticketmigration\ProfileRight::RIGHT_CONFIG, READ)
+    && !Session::haveRight('config', UPDATE)) {
+    Html::displayErrorAndDie(__('You do not have permission to perform this action.'));
+}
 Html::header(__('Ticket Migration configuration', 'ticketmigration'), $_SERVER['PHP_SELF'], 'tools', \GlpiPlugin\Ticketmigration\Menu::class);
 echo '<div class="alert alert-info m-3">' . __s('Plugin configuration is under development.', 'ticketmigration') . '</div>';
 Html::footer();